Explanation of code
Line 3 - convert a integer (whole number to floating number) 

Line 5 - 18 - A function which take an integer p and a list of all primes less then p and return a boolean value (true if p is a prime else false)

Line 6 - Empty list case
Line 7 - Test if the first item of the list is greater then the squareroot of the number testet. (round down)
Line 8  - case when p modulo x is not equal zero, recursively test the tail of the list (xs is the tail)
line  9  - Anything else test false (p modulo n = 0)

Line 11 - A function which take a integer n and returns a list of all primes less then or equal 2*n +1.
Line 12 - A predefined function which take a list an reverse it. (reverse the result of  line 13 - 18)
Line 13 - A predefined function which take a accumulator function an accumulator and a list to take the function over.
Line 14 - 17 - The accumulator function, adding the prime to the accumulation list (acc) if p is a prime and using acc to test against.
Line 18 - [] is the accumulator (empty list), (2 :: [for i in 1 .. n -> 2*i+1]) is 2 is the first number of the test list, and the rest are the n'th first odd numbers greater then 1.